feature: new moudle import mechanism.
The mechanism can import modules from a source file or directory. I
write a module loader for the mechanism, it can search modules from the
directory where the current running file is in or the CANDY_HOME/libs
(Environment Path).

If a directory is loaded, it first searchs the __module__.cd from the
directory. loader will run it if the file is existed; otherwise,
all the files with the suffix '.cd' in the directory will be loaded as
a module.